Rename `resource_manager` package to `manager`.
