+++ manager/doc.go
@@ -1,0 +1,2 @@
+// Package manager provides an API for loading and managing resources from a file system.
+package manager
