--- resource_manager/doc.go
@@ -1,2 +1,0 @@
-// Package resource_manager provides an API for loading and managing resources from a file system.
-package resource_manager
